requirements.txt
mlflow
dvc
pandas
numpy
scikit-learn
pyyaml
matplotlib
joblib

